# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MCMS API
  description: Medical Clinic Management System - REST API Documentation
  version: 1.0.0
  contact:
    name: MCMS Support
servers:
  - url: /api
    description: Default server
paths:
...
  /patients/{patientId}:
    get:
      tags: [Patients]
      summary: Show patient
      operationId: patients.show
      security: [{bearerAuth: []}]
      parameters:
        - name: patientId
          in: path
          required: true
          schema: {type: integer}
      responses:
        '200':
          description: Patient details
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Patient'}
    put:
      tags: [Patients]
      summary: Update patient
      operationId: patients.update
      security: [{bearerAuth: []}]
      parameters:
        - name: patientId
          in: path
          required: true
          schema: {type: integer}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name: {type: string}
                last_name: {type: string}
                date_of_birth: {type: string, format: date}
                gender: {type: string, enum: [male, female, other]}
                phone: {type: string}
                email: {type: string, format: email}
      responses:
        '200':
          description: Patient updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Patient'}
    delete:
      tags: [Patients]
      summary: Delete patient
      operationId: patients.destroy
      security: [{bearerAuth: []}]
      parameters:
        - name: patientId
          in: path
          required: true
          schema: {type: integer}
      responses:
        '200':
          description: Patient deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Patient:
      type: object
      properties:
        id: {type: integer}
        clinic_id: {type: integer}
        file_number: {type: string}
        first_name: {type: string}
        last_name: {type: string}
        date_of_birth: {type: string, format: date}
        gender: {type: string, enum: [male, female, other]}
        phone: {type: string}
        email: {type: string, format: email}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...